projects/insights-trends-dashboard/requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database Connectivity  
mysql-connector-python==8.2.0
